requirements.txt
#  CompKG-QA        AI Python dependencies
# Python: >=3.10,<3.13

# --- Web framework ---
fastapi
uvicorn
starlette

# --- LLM / Agent stack ---
langchain
langchain-core
langchain-community
langchain-openai
langchain-text-splitters
langgraph
langchain-tavily
langchain-mcp-adapters
langchain-deepseek
langchain-together
openai
faiss

# --- Knowledge base & graph ---
chromadb
lightrag-hku
neo4j
sqlalchemy

# --- Data & document processing ---
numpy
pandas
pymupdf
python-docx
markdownify
beautifulsoup4
pillow
pypinyin

# --- OCR / CV ---
rapidocr-onnxruntime
#cn-clip
torch
mineru

# --- Storage / auth / network ---
minio
pymysql
python-jose
PyJWT
python-dotenv
httpx
requests

# --- Utilities ---
loguru
pyyaml
tenacity
tqdm
typer
###
# --- Web framework ---
fastapi==0.115.12
uvicorn[standard]==0.34.0
starlette==0.46.2

# --- LLM / Agent stack ---
langchain==0.3.23
langchain-core==0.3.51
langchain-community==0.3.21
langchain-openai==0.3.12
langchain-text-splitters==0.3.8
langgraph==0.3.20
langchain-tavily==0.1.6
langchain-mcp-adapters==0.0.9
langchain-deepseek==0.1.3
langchain-together==0.3.0
openai==1.72.0
faiss-cpu==1.10.0

# --- Knowledge base & graph ---
chromadb==0.6.3
lightrag-hku==1.4.6
neo4j==5.28.1
sqlalchemy==2.0.40

# --- Data & document processing ---
numpy==1.26.4
pandas==2.2.3
pymupdf==1.25.5
python-docx==1.1.2
markdownify==1.1.0
beautifulsoup4==4.13.3
pillow==10.4.0
pypinyin==0.53.0

# --- OCR / CV ---
rapidocr-onnxruntime==1.4.4
cn-clip
torch==2.6.0
mineru==2.1.6

# --- Storage / auth / network ---
minio==7.2.15
pymysql==1.1.1
python-jose==3.3.0
PyJWT==2.10.1
python-dotenv==1.1.0
httpx==0.28.1
requests==2.32.3

# --- Utilities ---
loguru==0.7.3
pyyaml==6.0.2
tenacity==9.0.0
tqdm==4.67.1
typer==0.15.2